+++ internal/terminalui/terminal_other.go
@@ -19,6 +19,8 @@ func terminalWidth(_ *os.File) int {
 	return 0
 }
 
+func terminalHeight(_ *os.File) int { return 0 }
+
 func drawConsoleStatus(_ *os.File, _ string) bool { return false }
 
 func clearConsoleStatus(_ *os.File) bool { return false }